+++ astarte-dashboard/cypress/integration/device_page_test.js
@@ -53,10 +53,10 @@ describe('Device page tests', () => {
             cy.contains('Add new alias').should('exist').and('not.be.disabled');
           });
 
-        cy.contains('Metadata')
+        cy.contains('Attributes')
           .next()
           .within(() => {
-            cy.contains('Device has no metadata');
+            cy.contains('Device has no attribute');
             cy.contains('Add new item').should('exist').and('not.be.disabled');
           });
 
@@ -119,12 +119,12 @@ describe('Device page tests', () => {
             cy.contains('Add new alias').should('not.be.disabled');
           });
 
-        cy.contains('Metadata')
+        cy.contains('Attributes')
           .next()
           .within(() => {
-            Object.entries(this.deviceDetailed.data.metadata).forEach(
-              ([metadataKey, metadataValue]) => {
-                cy.contains(metadataKey).next().contains(metadataValue);
+            Object.entries(this.deviceDetailed.data.attributes).forEach(
+              ([attributeKey, attributeValue]) => {
+                cy.contains(attributeKey).next().contains(attributeValue);
               },
             );
             cy.contains('Add new item').should('not.be.disabled');
@@ -395,16 +395,16 @@ describe('Device page tests', () => {
       });
     });
 
-    it('correctly adds a device metadata', function () {
+    it('correctly adds a device attribute', function () {
       const device = _.merge({}, this.device);
-      device.data.metadata = {};
+      device.data.attributes = {};
       const updatedDevice = _.merge({}, this.device);
-      updatedDevice.data.metadata = { metadata_key: 'metadata_value' };
+      updatedDevice.data.attributes = { attribute_key: 'attribute_value' };
       cy.intercept('GET', '/appengine/v1/*/devices/*', device);
       cy.visit(`/devices/${device.data.id}/edit`);
       cy.get('.main-content').within(() => {
         cy.get('.card-header')
-          .contains('Metadata')
+          .contains('Attributes')
           .parents('.card')
           .within(() => {
             cy.contains('Add new item').should('exist').and('not.be.disabled').click();
@@ -414,9 +414,9 @@ describe('Device page tests', () => {
           .parents('.modal')
           .within(() => {
             cy.get('button').contains('Confirm').should('be.disabled');
-            cy.get('input#key').paste('metadata_key');
+            cy.get('input#key').paste('attribute_key');
             cy.get('button').contains('Confirm').should('not.be.disabled');
-            cy.get('input#value').paste('metadata_value');
+            cy.get('input#value').paste('attribute_value');
             cy.intercept('GET', '/appengine/v1/*/devices/*', updatedDevice);
             cy.intercept('PATCH', '/appengine/v1/*/devices/*', updatedDevice).as(
               'updateDeviceRequest',
@@ -426,42 +426,42 @@ describe('Device page tests', () => {
         cy.wait('@updateDeviceRequest')
           .its('request.body')
           .then((body) => JSON.parse(body).data)
-          .should('deep.eq', { metadata: { metadata_key: 'metadata_value' } });
+          .should('deep.eq', { attributes: { attribute_key: 'attribute_value' } });
         cy.get('.card-header')
-          .contains('Metadata')
+          .contains('Attributes')
           .parents('.card')
           .within(() => {
             cy.get('table tbody').find('tr').should('have.length', 1);
-            cy.contains('metadata_key');
-            cy.contains('metadata_value');
+            cy.contains('attribute_key');
+            cy.contains('attribute_value');
           });
       });
     });
 
-    it('correctly removes a device metadata', function () {
+    it('correctly removes a device attribute', function () {
       const device = _.merge({}, this.device);
-      device.data.metadata = { metadata_key1: 'metadata_value1', metadata_key2: 'metadata_value2' };
+      device.data.attributes = { attribute_key1: 'attribute_value1', attribute_key2: 'attribute_value2' };
       const updatedDevice = _.merge({}, this.device);
-      updatedDevice.data.metadata = { metadata_key1: 'metadata_value1' };
+      updatedDevice.data.attributes = { attribute_key1: 'attribute_value1' };
       cy.intercept('GET', '/appengine/v1/*/devices/*', device);
       cy.visit(`/devices/${device.data.id}/edit`);
       cy.get('.main-content').within(() => {
         cy.get('.card-header')
-          .contains('Metadata')
+          .contains('Attributes')
           .parents('.card')
           .within(() => {
             cy.get('table tbody').find('tr').should('have.length', 2);
-            cy.contains('metadata_key1');
-            cy.contains('metadata_value1');
-            cy.contains('metadata_key2');
-            cy.contains('metadata_value2');
+            cy.contains('attribute_key1');
+            cy.contains('attribute_value1');
+            cy.contains('attribute_key2');
+            cy.contains('attribute_value2');
             cy.get('table tbody tr:nth(1) i.fa-eraser').click();
           });
         cy.get('.modal-header')
           .contains('Delete Item')
           .parents('.modal')
           .within(() => {
-            cy.contains('Do you want to delete metadata_key2 from metadata?');
+            cy.contains('Do you want to delete attribute_key2 from attributes?');
             cy.intercept('GET', '/appengine/v1/*/devices/*', updatedDevice);
             cy.intercept('PATCH', '/appengine/v1/*/devices/*', updatedDevice).as(
               'updateDeviceRequest',
@@ -471,42 +471,42 @@ describe('Device page tests', () => {
         cy.wait('@updateDeviceRequest')
           .its('request.body')
           .then((body) => JSON.parse(body).data)
-          .should('deep.eq', { metadata: { metadata_key2: null } });
+          .should('deep.eq', { attributes: { attribute_key2: null } });
         cy.get('.card-header')
-          .contains('Metadata')
+          .contains('Attributes')
           .parents('.card')
           .within(() => {
             cy.get('table tbody').find('tr').should('have.length', 1);
-            cy.contains('metadata_key1');
-            cy.contains('metadata_value1');
+            cy.contains('attribute_key1');
+            cy.contains('attribute_value1');
           });
       });
     });
 
-    it('correctly edits a device metadata', function () {
+    it('correctly edits a device attribute', function () {
       const device = _.merge({}, this.device);
-      device.data.metadata = { metadata_key: 'metadata_value' };
+      device.data.attributes = { attribute_key: 'attribute_value' };
       const updatedDevice = _.merge({}, this.device);
-      updatedDevice.data.metadata = { metadata_key: 'metadata_new_value' };
+      updatedDevice.data.attributes = { attribute_key: 'attribute_new_value' };
       cy.intercept('GET', '/appengine/v1/*/devices/*', device);
       cy.visit(`/devices/${device.data.id}/edit`);
       cy.get('.main-content').within(() => {
         cy.get('.card-header')
-          .contains('Metadata')
+          .contains('Attributes')
           .parents('.card')
           .within(() => {
             cy.get('table tbody').find('tr').should('have.length', 1);
-            cy.contains('metadata_key');
-            cy.contains('metadata_value');
+            cy.contains('attribute_key');
+            cy.contains('attribute_value');
             cy.get('table tbody tr:nth(0) i.fa-pencil-alt').click();
           });
         cy.get('.modal-header')
-          .contains('Edit "metadata_key"')
+          .contains('Edit "attribute_key"')
           .parents('.modal')
           .within(() => {
             cy.get('input#value').clear();
             cy.get('button').contains('Confirm').should('not.be.disabled');
-            cy.get('input#value').paste('metadata_new_value');
+            cy.get('input#value').paste('attribute_new_value');
             cy.intercept('GET', '/appengine/v1/*/devices/*', updatedDevice);
             cy.intercept('PATCH', '/appengine/v1/*/devices/*', updatedDevice).as(
               'updateDeviceRequest',
@@ -516,14 +516,14 @@ describe('Device page tests', () => {
         cy.wait('@updateDeviceRequest')
           .its('request.body')
           .then((body) => JSON.parse(body).data)
-          .should('deep.eq', { metadata: { metadata_key: 'metadata_new_value' } });
+          .should('deep.eq', { attributes: { attribute_key: 'attribute_new_value' } });
         cy.get('.card-header')
-          .contains('Metadata')
+          .contains('Attributes')
           .parents('.card')
           .within(() => {
             cy.get('table tbody').find('tr').should('have.length', 1);
-            cy.contains('metadata_key');
-            cy.contains('metadata_new_value');
+            cy.contains('attribute_key');
+            cy.contains('attribute_new_value');
           });
       });
     });